Security update, fix CVE-2006-5793, details follows:
Tavis Ormandy has reported a vulnerability in libpng, which can be exploited by malicious people to cause a DoS (Denial of Service).
 
The vulnerability is caused due to an out-of-bounds read error in the "png_set_sPLT()" function in pngset.c. This can be exploited by tricking an 
application using the library to process a specially crafted PNG file.