Licence change,
Hi Folks:

At the request of the open-source community, the JasPer software has
been re-released under a new less-restrictive license. The new license
is a truly open-source license and is based on the popular MIT
license. The release with the new license can be downloaded from the
JasPer Project Home Page.

Regards,
Michael