Beagle 0.2.15 is out. As always, bug fixes, performance improvements, new features. We’ve got it all.
This version should be much better behaved than previous versions. We now take advantage of Linux’s SCHED_BATCH scheduling policy when available, we’ve replaced the RTF 
filter, we’ve moved the troublesome MS Word filter out of process for reliability reasons, and we’ve fixed problems with looping directories.

We have new KDE backends: KOrganizer and Konversation. I’m hoping the Konversation backend can be used as a template for things like an X-Chat backend (hint, hint).

And for the Mono geeks in the audience, this release is the first following our switch to Mono’s gmcs compiler (finally). We’re now on the .NET 2.0 profile, which has been 
a focus of the Mono community for some time. The class libraries are improved, but the biggest win for us is generics. We now get typesafe collections and value types are 
no longer boxed, saving us a bunch of unnecessary allocations. Hooray.