dhcpcd-3.0.11 Don't try to send messages when we infinite leasetime and the interface fd is closed.
dhcpcd-3.0.10
Only write a new ntp.conf if any of our servers are not present in it.
We now work with SIGCHLD and call wait so that we don't leave any
zombies lying around.
For infinite timeout, we now resent the last request at +TIMEOUT_MINI
intervals until TIMEOUT_MINI_INF is reached, thanks to siadak.
We now return a non zero exit code on SIGTERM and SIGINT if we have not
forked into the background.
When NIS and/or NTP servers are updated, we restart the service for them
if we can.

dhcpcd-3.0.9
Add static routes before any default routes as a router may require a host
route in the static routes.