This fixes the wrong master channel selection in 4 XYZ machines in our lab at default installs.
The problem is really this:
http://bugs.kde.org/show_bug.cgi?id=136416

This fix is a partial solution but a working one (at least for these machines 
in the lab and should also be fixing more machines that has channels named "front" 
and no "master")

It is partial because in KDE4 it is always PCM:0 that is set when it exists, other 
than that, it is "master", "front" or "pc speaker". But in KDE3 PCM has no mute 
option.