Initial release of the v4l-dvb package. This package contains the mercurial snapshot of http://linuxtv.org/hg which includes device drivers for various video capture devices, radio adapters and DVB adapters.
Current problems:
-> Either we have to delete the kernel modules already shipped
with our stock kernel package or install the modules shipped
with this package in extra/ directory and modify depmod.conf
in postInstall to change the subdir priority in /lib/modules.

-> All the drivers depending on SND are disabled because
we don't enable ALSA in our stock kernel.

Good news:
-> It compiles fine, the source tree is well maintained,
the drivers known to not function with the current kernel
are automatically disabled during compilation.

Hope this package will help people having DVB adapters.