Merge from devel/server/samba/:
rev. 77669, by bahadir.kandemir on 2009-10-01
Fixed winbind service not working

rev. 77691, by ozan.caglayan on 2009-10-02
CVE-2009-2813:
In all versions of Samba later than 3.0.11, connecting to the home
share of a user will use the root of the filesystem
as the home directory if this user is misconfigured to have
an empty home directory in /etc/passwd.

CVE-2009-2948:
If mount.cifs is installed as a setuid program, a user can pass it a
credential or password path to which he or she does not have access and
then use the --verbose option to view the first line of that file.
All known Samba versions are affected.

CVE-2009-2906:
Specially crafted SMB requests on authenticated SMB connections can
send smbd into a 100% CPU loop, causing a DoS on the Samba server.

--

Version bump to fix CVE-2009-2813, CVE-2009-2948, CVE-2009-2906 (#11250).

BUG:COMMENT:11250