Merge from devel/multimedia/graphics/gimp:
rev. 82744, by onur on 2009-12-14
forcely depend on new cairo, I am doing this only in devel so that it will be efective when packages are merged

If you are really really sure your package works with older cairo correctly you can remove the dependency

rev. 84713, by ozan.caglayan on 2010-01-14
Bump to 2.6.8