Merge from devel/network/plugin/flashplugin:
rev. 82742, by onur on 2009-12-14
forcely depend on new gtk, I am doing this only in devel so that it will be efective when packages are merged

If you are really really sure your package works with older gtk2 correctly you can remove the dependency

rev. 83470, by inanc on 2009-12-24
Version bump to new beta release

rev. 87090, by erenturkay on 2010-02-17
Update to 10.0.45.2 to fix multiple vulnerabilities (CVE-2010-018{6,7})

BUG:COMMENT:12309

rev. 87091, by erenturkay on 2010-02-17
since this is binary library, it is not required to specify gtk dependency version