Merge from devel/game/fps/alienarena:
rev. 90316, by igungor on 2010-04-19
Change mail address a.k.a. "Packagers email address is wrong"

rev. 90691, by igungor on 2010-04-29
Fix library search path.

I could have sworn that i played this game
for about an hour and worked like a charm.
And now, it complains about that no library is found and etc.