pisi 2.4_alpha3
This is the third alpha for the upcoming 2.4.x releases. This release
introduces a new feature for building source packages with different
configurations using the same actions.py. This is achieved with
BuildType tags specified between Package tags.

It is also possible to write build dependencies inside Package tags with
this release. This is needed since additional build dependencies could
be needed for different build types.

Changes since 2.4_alpha2
------------------------
fatih (12):
      build: Minor simplifications
      build: More simplifications
      build: Cosmetic change
      build: Check build dependencies for unpack too
      sourcearchive: Pass target dirs to unpack methods
      specfile: Accept BuildType tag inside Package tags
      build: Add support for build types
      build: Colorize unpack message
      specfile: Accept BuildDependencies inside Package tags
      build: Iterate over Package tags to gather all build deps
      build: Remove redundant check in previous commit
      build,config: Add a new config option to ignore build types

ozan.caglayan (1):
      actionsapi/kerneltools: Don't fail if listnewconfig is not available

transifex (4):
      l10n: Updated Polish (pl) translation to 46% (Michał Komasa)
      l10n: Updated Polish (pl) translation to 47% (Michał Komasa)
      l10n: add uk (Коростіль Данило)
      l10n: Updated Italian (it) translation to 81% (Ali Işıngör)