pisi 2.4_rc1
This is the first release candidate for the upcoming 2.4.x releases.
This release fixes a regression introduced in beta2 and fixes a cosmetic
issue in the output of index command.

I think Pisi is quite ready for a stable release. Therefore, this may be
the last release before 2.4.

Changes since 2.4_beta2
-----------------------
fatih (4):
      MANIFEST: Add pisi.xml.in to the tarball
      Revert "build: Remove redundant dict variable"
      build: Add a comment for the dict variable
      index: Print CR character at the end