Clementine tarball from the git hash: 9dd45dbe82b2f57f8f7c4a1201b79b2c19e513a2 The tarball works fine with iPhone/iPod devices (reading or writing on disk) However the commit after, which has the hash: ccb9f8cf940939d9a8d146b8cd6d777086bb39ed breaks things. It's not possible to listen any song on these devies. This problem is still not fixed in the master branch, that means our package in devel repo is broken too.
The patch included in this source package is the same as the commit object above. Applying the patch
will create the broken package, not applying will create the working package.

Now we have to examinate the patch and find out where exactly the problem is. My first guess
is that the files:

    src/devices/gpoddevice.cpp
    src/devices/gpodloader.cpp
    src/devices/imobiledeviceconnection.cpp

are the first places where we should look for.