Merge from devel/programming/scm/git:
rev. 122989, by fatih.arslan on 2011-09-08
Dont build it for now, there is a problem with 1.7.6.1. The problem is fixed
with 1.7.6.2, however the tarball on kernel.org is still not available due the
shortages. We need to wait until it is available public. (The git repo itself
is also unavailable to make a patch on top of 1.7.6.1. Murphy, we love you)

More info for the "git push --quiet" bug:

    http://thread.gmane.org/gmane.comp.version-control.git/180652