Merge from devel/programming/language/python/python-mygpoclient:
rev. 119444, by igungor on 2011-05-27
Sync translations.

rev. 123230, by sdalgic on 2011-09-16
Version bump: This release is a catch-up release with various changes that accumulated in the repository over the last few months. This release includes experimental support for the mygpo-feedservice. It also includes various fixes to the gpodder.net API client, so it is a recommended release for all users of mygpoclient.