add gimp-wavelet-decompose to pg
This plugin losslessly decomposes a layer of an image into layers of wavelet scales.
This means that you can edit the image on different detail scales (frequencies).
The trivial recomposition of the image can be done by GIMP's 
layer modes so you can see the results of your modifications instantly.
Among the applications are retouching, noise reduction, and enhancing global contrast.

http://registry.gimp.org/node/11742