add gimp-wavelet-denoise to pg
The wavelet denoise plugin is a tool to reduce noise in each channel of an image separately.
The default colour space to do denoising is YCbCr which has the advantage that chroma noise
can be reduced without affecting image details. Denoising in CIELAB (L*a*b*) or RGB
is available as an option. The user interface allows colour mode and 
preview channel selection. The denoising threshold can be set for
each colour channel independently.

http://registry.gimp.org/node/4235